# frontend/默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
    "/comments": {
      "get": {
        "summary": "前台评论列表（仅返回状态为 APPROVED 的评论，树形结构）。",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "articleId",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListCommentTreeVO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": [
                    {
                      "id": 0,
                      "userId": 0,
                      "nickname": "",
                      "avatar": "",
                      "content": "",
                      "createTime": "",
                      "likeCount": 0,
                      "children": [
                        {
                          "id": 0,
                          "userId": 0,
                          "nickname": "",
                          "avatar": "",
                          "content": "",
                          "createTime": "",
                          "likeCount": 0,
                          "children": []
                        }
                      ]
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "新增评论（要求登录）。",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreateDTO",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLong"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": 0
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/comments/{id}/reply": {
      "post": {
        "summary": "回复某条评论（要求登录）。",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentReplyDTO",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLong"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": 0
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/comments/{id}/like": {
      "post": {
        "summary": "点赞评论（要求登录）。",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/comments/emoji/preview": {
      "put": {
        "summary": "previewEmoji",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentEmojiDTO",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": ""
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/orders": {
      "post": {
        "summary": "create",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreateDTO",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderVO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "orderNo": "",
                    "userId": 0,
                    "articleId": 0,
                    "amount": 0,
                    "status": "",
                    "payChannel": "",
                    "payTime": "",
                    "createTime": "",
                    "expireTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "get": {
        "summary": "allOrders",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIPageOrderVO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "orderNo": "",
                        "userId": 0,
                        "articleId": 0,
                        "amount": 0,
                        "status": "",
                        "payChannel": "",
                        "payTime": "",
                        "createTime": "",
                        "expireTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": "",
                    "pages": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/orders/{orderNo}/pay-success": {
      "post": {
        "summary": "paySuccess",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "orderNo",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/orders/{orderNo}/pay": {
      "post": {
        "summary": "createPayment",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "orderNo",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderPayRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseOrderPaymentVO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": {
                    "orderNo": "",
                    "amount": 0,
                    "payChannel": "",
                    "status": "",
                    "payUrl": "",
                    "expireTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/orders/{orderNo}/pay/balance": {
      "post": {
        "summary": "payByBalance",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "orderNo",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/orders/{orderNo}/mock-pay": {
      "post": {
        "summary": "mockPay",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "orderNo",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/orders/my": {
      "get": {
        "summary": "myOrders",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "pageNum",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseIPageOrderVO"
                },
                "example": {
                  "success": false,
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "orderNo": "",
                        "userId": 0,
                        "articleId": 0,
                        "amount": 0,
                        "status": "",
                        "payChannel": "",
                        "payTime": "",
                        "createTime": "",
                        "expireTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": "",
                    "pages": 0
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "CommentTreeVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "nickname": {
            "type": "string",
            "description": ""
          },
          "avatar": {
            "type": "string",
            "description": ""
          },
          "content": {
            "type": "string",
            "description": ""
          },
          "createTime": {
            "type": "string",
            "description": ""
          },
          "likeCount": {
            "type": "integer",
            "description": ""
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentTreeVO",
              "description": "评论树节点。"
            },
            "description": "",
            "default": "new ArrayList<>()"
          }
        }
      },
      "ApiResponseListCommentTreeVO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentTreeVO",
              "description": "评论树节点。"
            },
            "description": "具体数据内容"
          }
        }
      },
      "ApiResponseLong": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "type": "integer",
            "description": "具体数据内容",
            "format": "int64"
          }
        }
      },
      "CommentCreateDTO": {
        "type": "object",
        "properties": {
          "articleId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "parentId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "content": {
            "type": "string",
            "description": "",
            "maxLength": 1000
          }
        },
        "required": [
          "articleId",
          "content"
        ]
      },
      "CommentReplyDTO": {
        "type": "object",
        "properties": {
          "articleId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "content": {
            "type": "string",
            "description": "",
            "maxLength": 1000
          }
        },
        "required": [
          "articleId",
          "content"
        ]
      },
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "description": "具体数据内容",
            "type": "null"
          }
        }
      },
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "type": "string",
            "description": "具体数据内容"
          }
        }
      },
      "CommentEmojiDTO": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": ""
          }
        },
        "required": [
          "content"
        ]
      },
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "column": {
            "type": "string",
            "description": ""
          },
          "asc": {
            "type": "boolean",
            "description": ""
          }
        }
      },
...
      "OrderVO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orderNo": {
            "type": "string",
            "description": ""
          },
          "userId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "articleId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "amount": {
            "type": "number",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": ""
          },
          "payChannel": {
            "type": "string",
            "description": ""
          },
          "payTime": {
            "type": "string",
            "description": ""
          },
          "createTime": {
            "type": "string",
            "description": ""
          },
          "expireTime": {
            "type": "string",
            "description": ""
          }
        }
      },
      "ApiResponseOrderVO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "$ref": "#/components/schemas/OrderVO",
            "description": "具体数据内容"
          }
        }
      },
      "OrderCreateDTO": {
        "type": "object",
        "properties": {
          "articleId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "amount": {
            "type": "number",
            "description": ""
          },
          "payChannel": {
            "type": "string",
            "description": ""
          }
        },
        "required": [
          "articleId",
          "amount"
        ]
      },
      "OrderPaymentVO": {
        "type": "object",
        "properties": {
          "orderNo": {
            "type": "string",
            "description": ""
          },
          "amount": {
            "type": "number",
            "description": ""
          },
          "payChannel": {
            "type": "string",
            "description": ""
          },
          "status": {
            "type": "string",
            "description": ""
          },
          "payUrl": {
            "type": "string",
            "description": ""
          },
          "expireTime": {
            "type": "string",
            "description": ""
          }
        }
      },
      "ApiResponseOrderPaymentVO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "$ref": "#/components/schemas/OrderPaymentVO",
            "description": "具体数据内容"
          }
        }
      },
      "OrderPayRequest": {
        "type": "object",
        "properties": {
          "payChannel": {
            "type": "string",
            "description": ""
          }
        },
        "required": [
          "payChannel"
        ]
      },
      "IPageOrderVO": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderVO",
              "description": "订单视图对象。"
            },
            "description": ""
          },
          "total": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "current": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem",
              "description": "com.baomidou.mybatisplus.core.metadata.OrderItem"
            },
            "description": ""
          },
          "optimizeCountSql": {
            "type": "boolean",
            "description": ""
          },
          "searchCount": {
            "type": "boolean",
            "description": ""
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "description": ""
          },
          "maxLimit": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "countId": {
            "type": "string",
            "description": ""
          },
          "pages": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ApiResponseIPageOrderVO": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "是否成功"
          },
          "code": {
            "type": "integer",
            "description": "状态码"
          },
          "message": {
            "type": "string",
            "description": "文案信息（给前端使用）"
          },
          "data": {
            "$ref": "#/components/schemas/IPageOrderVO",
            "description": "具体数据内容"
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}